pyproject.toml
[tool.poetry]
name = "wolfery-log-filter"
version = "0.1.0"
description = ""
authors = ["Madison Scott-Clary <makyo@drab-makyo.com>"]
readme = "README.md"
packages = [{include = "wolfery_log_filter"}]

[tool.poetry.dependencies]
python = "^3.10"
beautifulsoup4 = "^4.12.2"
ipython = "^8.14.0"
pyyaml = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
